requirements.txt
opencv-python==4.5.5.64
Pillow==9.2.0
numpy==1.23.3
ultralytics==0.1.0